requirement.txt
streamlit==1.24.0
cv2
numpy
os
pillow
imutils
keras
